pyproject.toml
[project]
name = "alphavault"
version = "0.1.0"
description = "AlphaVault"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "feedparser",
  "litellm",
  "python-dotenv",
  "redis",
  "requests",
  "sqlalchemy",
  "sqlalchemy-libsql",
  "streamlit",
  "supervisor",
]

[dependency-groups]
dev = [
  "pre-commit>=3.0.0",
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
